requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile
#
amqp==5.3.1
    # via
    #   -r requirements.in
    #   kombu
asgiref==3.8.1
    # via
    #   -r requirements.in
    #   django
billiard==4.2.1
    # via
    #   -r requirements.in
    #   celery
build==1.2.2.post1
    # via
    #   -r requirements.in
    #   pip-tools
celery==5.5.1
    # via -r requirements.in
click==8.1.8
    # via
    #   -r requirements.in
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
    #   pip-tools
    #   typer
click-didyoumean==0.3.1
    # via
    #   -r requirements.in
    #   celery
click-plugins==1.1.1
    # via
    #   -r requirements.in
    #   celery
click-repl==0.3.0
    # via
    #   -r requirements.in
    #   celery
dj-database-url==2.3.0
    # via -r requirements.in
django==5.2
    # via
    #   -r requirements.in
    #   dj-database-url
    #   djangorestframework
djangorestframework==3.16.0
    # via -r requirements.in
kombu==5.5.3
    # via
    #   -r requirements.in
    #   celery
markdown-it-py==3.0.0
    # via rich
mdurl==0.1.2
    # via markdown-it-py
packaging==25.0
    # via
    #   -r requirements.in
    #   build
pip-tools==7.4.1
    # via -r requirements.in
prompt-toolkit==3.0.51
    # via
    #   -r requirements.in
    #   click-repl
psycopg2==2.9.10
    # via -r requirements.in
pygments==2.19.1
    # via rich
pyproject-hooks==1.2.0
    # via
    #   -r requirements.in
    #   build
    #   pip-tools
python-dateutil==2.9.0.post0
    # via
    #   -r requirements.in
    #   celery
python-decouple==3.8
    # via -r requirements.in
redis==5.2.1
    # via -r requirements.in
rich==14.0.0
    # via typer
shellingham==1.5.4
    # via typer
six==1.17.0
    # via
    #   -r requirements.in
    #   python-dateutil
sqlparse==0.5.3
    # via
    #   -r requirements.in
    #   django
typer==0.15.2
    # via -r requirements.in
typing-extensions==4.13.2
    # via
    #   -r requirements.in
    #   dj-database-url
    #   typer
tzdata==2025.2
    # via
    #   -r requirements.in
    #   kombu
vine==5.1.0
    # via
    #   -r requirements.in
    #   amqp
    #   celery
    #   kombu
wcwidth==0.2.13
    # via
    #   -r requirements.in
    #   prompt-toolkit
wheel==0.45.1
    # via
    #   -r requirements.in
    #   pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
